backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
httpx
aiohttp==3.9.5
websockets==12.0